RCJ_pcms_base/requirements.txt
tensorflow==1.14.0
tensorflow-gpu==1.14.0
keras==2.2.5
dlib>=19.16.0
opencv-python==4.1.1.26
opencv-contrib-python==4.1.1.26
numpy==1.17.3
imutils>=0.5.3
pyttsx3==2.71
dataclasses>=0.7
matplotlib>=3.2.1
Werkzeug~=1.0.0
rospy~=1.14.5
Pillow~=5.1.0
h5py~=2.10.0
requests~=2.23.0
rospkg~=1.2.4
actionlib~=1.12.0